完善用户请求方法

+++ src/main/java/com/chinamobile/iot/lightapp/mysql/service/Impl/WorkCycleServiceImpl.java
@@ -1,9 +1,12 @@
 package com.chinamobile.iot.lightapp.mysql.service.Impl;
 
 
+import com.chinamobile.iot.lightapp.mysql.config.Constant;
+import com.chinamobile.iot.lightapp.mysql.dao.RegionMapper;
+import com.chinamobile.iot.lightapp.mysql.dao.UserWorkcycleMapper;
 import com.chinamobile.iot.lightapp.mysql.dao.WorkCycleMapper;
-import com.chinamobile.iot.lightapp.mysql.model.WorkCycle;
-import com.chinamobile.iot.lightapp.mysql.model.WorkCycleExample;
+import com.chinamobile.iot.lightapp.mysql.model.*;
+import com.chinamobile.iot.lightapp.mysql.service.UserWorkcycleService;
 import com.chinamobile.iot.lightapp.mysql.service.WorkCycleService;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
@@ -24,15 +27,18 @@ public class WorkCycleServiceImpl implements WorkCycleService {
 
     @Autowired
     private WorkCycleMapper workCycleMapper;
+    @Autowired
+    private RegionMapper regionMapper;
+    @Autowired
+    private UserWorkcycleMapper userWorkcycleMapper;
 
     @Override
     public PageInfo<WorkCycle> findWorkCycles(WorkCycle workCycle, Integer pageNum, Integer pageSize) {
         WorkCycleExample workCycleExample = new WorkCycleExample();
         WorkCycleExample.Criteria criteria = workCycleExample.createCriteria();
-
-
-        PageHelper.startPage(pageNum, pageSize,true,false);
-        List<WorkCycle> list= workCycleMapper.selectByExample(workCycleExample);
+        criteria.andWorkCycleNameEqualTo(workCycle.getWorkCycleName());
+        PageHelper.startPage(pageNum, pageSize, true, false);
+        List<WorkCycle> list = workCycleMapper.selectByExample(workCycleExample);
         return new PageInfo<WorkCycle>(list);
     }
 
@@ -42,18 +48,54 @@ public class WorkCycleServiceImpl implements WorkCycleService {
     }
 
     @Override
-    public int deleteByWorkCycleId(Integer workCycleId) {
-        return workCycleMapper.deleteByPrimaryKey(workCycleId);
+    public int deleteByWorkCycleId(Integer userId, Integer workCycleId) {
+        UserWorkcycleExample userWorkcycleExample = new UserWorkcycleExample();
+        UserWorkcycleExample.Criteria criteria = userWorkcycleExample.createCriteria();
+        criteria.andWorkCycleIdEqualTo(workCycleId);
+        criteria.andUserIdEqualTo(userId);
+        criteria.andIsManagerEqualTo(Constant.CYCLE_CREATER);
+        List<UserWorkcycle> list = userWorkcycleMapper.selectByExample(userWorkcycleExample);
+        if (list == null && list.size() == 0) {
+            return NO_PERMISSION;
+        }
+        workCycleMapper.deleteByPrimaryKey(workCycleId);
+        return SUCCESS;
     }
 
     @Override
-    public int updateByWorkCycleId(WorkCycle workCycle) {
-        return workCycleMapper.updateByPrimaryKeySelective(workCycle);
+    public int updateByWorkCycleId(Integer userId, WorkCycle workCycle) {
+        UserWorkcycleExample userWorkcycleExample = new UserWorkcycleExample();
+        UserWorkcycleExample.Criteria criteria = userWorkcycleExample.createCriteria();
+        criteria.andWorkCycleIdEqualTo(workCycle.getWorkCycleId());
+        criteria.andUserIdEqualTo(userId);
+        criteria.andIsManagerEqualTo(Constant.CYCLE_CREATER);
+        List<UserWorkcycle> list = userWorkcycleMapper.selectByExample(userWorkcycleExample);
+        if (list == null && list.size() == 0) {
+            return NO_PERMISSION;
+        }
+        workCycleMapper.updateByPrimaryKeySelective(workCycle);
+        return SUCCESS;
     }
 
     @Override
-    public int insert(WorkCycle workCycle) {
-        return workCycleMapper.insertSelective(workCycle);
+    public int insert(Integer userId, WorkCycle workCycle, List<Region> regionList) {
+        //新增工作圈信息
+        workCycleMapper.insertSelective(workCycle);
+        //新增区域信息
+        Integer workCycleId = workCycle.getWorkCycleId();
+        if (regionList != null) {
+            for (Region temp : regionList) {
+                temp.setCycleId(workCycleId);
+                regionMapper.insert(temp);
+            }
+        }
+        //新增用户工作圈权限
+        UserWorkcycle userWorkcycle = new UserWorkcycle();
+        userWorkcycle.setUserId(userId);
+        userWorkcycle.setWorkCycleId(workCycleId);
+        userWorkcycle.setIsManager(Constant.CYCLE_CREATER);
+        userWorkcycleMapper.insert(userWorkcycle);
+        return workCycleId;
     }
 
 }
